requirements.txt
asgiref==3.9.2
diff-match-patch==20241021
Django==5.2.7
django-import-export==4.3.10
django-unfold==0.67.0
sqlparse==0.5.3
tablib==3.8.0
